Extract generic Lucene querying code into LuceneIndexServiceBase. The plan is to move that to nexus-indexer-data, see https://github.com/windup/nexus-repository-indexer/pull/5
